requirements.txt
ufal.udpipe
prettytable
tqdm
conllu
pandas
numpy